requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
flake8
python-dotenv>=0.5.1
pandas
numpy
sklearn
matplotlib
seaborn
jupyter
black
sqlalchemy
pymysql
statsmodels
pmdarima
joblib
holidays
tqdm
